requirements.txt
playwright==1.49.0
pytest==8.3.4
pytest-playwright==0.6.2
pytest-html==4.1.1
pytest-xdist==3.6.1
python-dotenv==1.0.1
allure-pytest==2.13.5
anthropic>=0.68.0
